webpack configuration
Foi realizada a instalação e configuração do "Webpack" para definir variáveis de ambiente, e assim tornar o site mais seguro. Protegendo os dados sensíveis.
