+++ .env
@@ -1,0 +1,1 @@
+API_KEY = "999c8ad61amshb2ce37ebbbd606bp161675jsn2de47b4f7943"
