+++ postcss.config.js
@@ -1,0 +1,7 @@
+const tailwindcss = require('tailwindcss');
+module.exports = {
+  plugins: [
+    'postcss-preset-env',
+    tailwindcss
+  ],
+};
