+++ src/ticket.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="/dist/output.css" rel="stylesheet" />
+    
 
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
@@ -18,7 +19,6 @@
       crossorigin="anonymous"
       referrerpolicy="no-referrer"
     />
-
     <script src="main.js" defer></script>
     <title>Document</title>
 </head>
