+++ webpack.config.js
@@ -1,0 +1,38 @@
+const path = require("path");
+
+const MiniCssExtractPlugin = require("mini-css-extract-plugin");
+
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+
+const Dotenv = require("dotenv-webpack");
+
+module.exports = {
+  mode: "development",
+  entry: "./src/main.js",
+  output: {
+    filename: "bundle.js",
+    path: path.resolve(__dirname, "./dist"),
+  },
+  devServer: {
+    static: path.resolve(__dirname, "dist"),
+    compress: true,
+    port: 9000,
+    hot: true,
+  },
+  module: {
+    rules: [
+      {
+        test: /\.css$/i,
+        include: path.resolve(__dirname, "src"),
+        use: ["style-loader", "css-loader", "postcss-loader"],
+      },
+    ],
+  },
+  plugins: [
+    new HtmlWebpackPlugin({
+      filename: "index.html",
+      template: "src/template.html",
+    }),
+    new Dotenv(),
+  ],
+};
